fix final

+++ lib/widgets/note_widget.dart
@@ -5,20 +5,46 @@ class NoteWidget extends StatelessWidget {
   final Note note;
   final Function onDelete;
   final Function onEdit;
+
   NoteWidget(
       {required this.note, required this.onDelete, required this.onEdit});
+
   @override
   Widget build(BuildContext context) {
-    return Dismissible(
-      key: Key(note.title),
-      background: Container(color: Colors.red),
-      onDismissed: (direction) {
-        onDelete();
+    return GestureDetector(
+      onLongPress: () {
+        showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+              title: Text('Pilih Aksi'),
+              actions: [
+                TextButton(
+                  onPressed: () {
+                    onEdit();
+                    // Navigator.of(context).pop();
+                  },
+                  child: Text('Edit'),
+                ),
+                TextButton(
+                  onPressed: () {
+                    onDelete();
+                    // Navigator.of(context).pop();
+                  },
+                  child: Text('Hapus'),
+                ),
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(),
+                  child: Text('Batal'),
+                ),
+              ],
+            );
+          },
+        );
       },
       child: ListTile(
         title: Text(note.title),
         subtitle: Text(note.content),
-        onLongPress: () => onEdit(),
       ),
     );
   }
